Sarah Kelsey Branch 2013.06.27
Decided to grab the photswipe responsive gallery for a base.
Added the json call with juery to load the images and image data.
Constructed the unordered list (required by photoswipe) than appended it
to the #Gallery element
Used Jquery mobile so that this will not only be responsive but will
also function across devices with swipe capabilities (was going to add
mouse wheel and keyboard arrows but thought that might be overkill.
Oh FYI the json file that was included in this project had some missing
and some improperly formatted data. There was an image that did not have
a title defined as well as a few that were not title case. I would have
put checks in the code to correct this but opted to change these items
in the json file itself as data errors really should be corrected at the
data source.
Enjoy =)